Issue #1: SectionFooterTitleType return Optional<String>
It's a common case when the same table contains sections with footer and without. SectionFooterTitleType return String. So the only option to add section without footer is return "", but it adds some extra redundant space between sections.
SectionFooterTitleType should return Optional<String> to mix sections with footer and without in the same table sections source.
https://github.com/bigMOTOR/DataDrivenRxTableView/issues/1